componentize nav menu
