+++ src/components/NavMenu/index.tsx
@@ -1,0 +1,1 @@
+export { NavMenu } from './NavMenu'
